requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp_socks==0.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bibtexparser==1.4.1
blinker==1.8.2
bottle==0.12.25
Brotli==1.1.0
browser-cookie3==0.19.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
citeproc-py==0.6.0
citeproc-py-styles==0.1.3
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
comm==0.2.2
cryptography==43.0.0
cssselect2==0.7.0
curl_cffi==0.7.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
duckduckgo_search==6.2.9
executing==2.0.1
faicons==0.2.2
fastapi==0.112.1
Flask==3.0.3
frozenlist==1.4.1
g4f==0.3.2.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmltools==0.5.3
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.3.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kaitaistruct==0.10
latexcodec==3.0.0
linkify-it-py==2.0.3
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
plyer==2.1.0
primp==0.6.0
prompt-toolkit==3.0.36
proxy_tools==0.1.0
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydivert==2.1.0
PyExecJS==1.5.1
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-socks==2.5.0
pythonnet==3.0.3
pytz==2024.1
pywebview==5.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.1
questionary==2.0.1
requests==2.32.3
retrying==1.3.4
selenium==4.23.1
selenium-wire==5.1.0
shiny==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.2
tenacity==9.0.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
undetected-chromedriver==3.5.5
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.23.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.0
zstandard==0.23.0
